# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 聆櫻聖境點歌系統 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /songs:
    get:
      operationId: listSongs
      tags: [songs]
      summary: List all songs with optional filters
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: language
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: primaryTag
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of songs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SongList"
    post:
      operationId: createSong
      tags: [songs]
      summary: Create a new song
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongInput"
      responses:
        "201":
          description: Song created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
  /songs/{id}:
    get:
      operationId: getSong
      tags: [songs]
      summary: Get a song by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Song detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
    patch:
      operationId: updateSong
      tags: [songs]
      summary: Update a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongUpdate"
      responses:
        "200":
          description: Song updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
    delete:
      operationId: deleteSong
      tags: [songs]
      summary: Delete a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Song deleted
  /songs/analyze-url:
    post:
      operationId: analyzeUrl
      tags: [songs]
      summary: Analyze a YouTube URL and suggest song metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/YouTubeUrlInput"
      responses:
        "200":
          description: Analyzed YouTube metadata with song suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YouTubeAnalysis"
  /songs/import/google-sheet/preview:
    post:
      operationId: previewGoogleSheetSync
      tags: [songs]
      summary: Preview what a Google Sheet sync would do
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleSheetImport"
      responses:
        "200":
          description: Sync preview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncPreview"
  /songs/import/google-sheet:
    post:
      operationId: importFromGoogleSheet
      tags: [songs]
      summary: Import songs from Google Sheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoogleSheetImport"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"
...
  /songs/{id}/cooldown:
    get:
      operationId: getSongCooldown
      tags: [songs]
      summary: Get cooldown status for a song
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cooldown status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CooldownStatus"
...
components:
  schemas:
...
    Song:
      type: object
      required: [id, title, artist, language, playCount, categories, youtubeUrl, isPracticing, hasPitchWarning, createdAt, cooldownDays, cooldownMode]
      properties:
        id:
          type: integer
        title:
          type: string
        artist:
          type: string
        language:
          type: string
        playCount:
          type: integer
        primaryTag:
          type: ["string", "null"]
        categories:
          type: array
          items:
            type: string
        youtubeUrl:
          type: ["string", "null"]
        isPracticing:
          type: boolean
        hasPitchWarning:
          type: boolean
        status:
          type: string
        createdAt:
          type: string
        lyricsText:
          type: ["string", "null"]
        lyricsFormat:
          type: ["string", "null"]
        lyricsSource:
          type: ["string", "null"]
        lyricsUpdatedAt:
          type: ["string", "null"]
        cooldownDays:
          type: integer
        cooldownMode:
          type: string
          enum: [warn, block]
    SongInput:
      type: object
      required: [title, artist, language]
      properties:
        title:
          type: string
        artist:
          type: string
        language:
          type: string
        status:
          type: string
        primaryTag:
          type: ["string", "null"]
        categories:
          type: array
          items:
            type: string
        youtubeUrl:
          type: ["string", "null"]
        isPracticing:
          type: boolean
        hasPitchWarning:
          type: boolean
        cooldownDays:
          type: integer
        cooldownMode:
          type: string
    SongUpdate:
      type: object
      properties:
        title:
          type: string
        artist:
          type: string
        language:
          type: string
        status:
          type: string
        primaryTag:
          type: ["string", "null"]
        categories:
          type: array
          items:
            type: string
        youtubeUrl:
          type: ["string", "null"]
        isPracticing:
          type: boolean
        hasPitchWarning:
          type: boolean
        cooldownDays:
          type: integer
        cooldownMode:
          type: string
    SongList:
      type: object
      required: [songs, total, page, limit]
      properties:
        songs:
          type: array
          items:
            $ref: "#/components/schemas/Song"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    CooldownStatus:
      type: object
      required: [cooldownDays, cooldownMode, isInCooldown]
      properties:
        cooldownDays:
          type: integer
        cooldownMode:
          type: string
          enum: [warn, block]
        lastPerformed:
          type: ["string", "null"]
        remainingDays:
          type: integer
        isInCooldown:
          type: boolean
...
    YouTubeUrlInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
    YouTubeAnalysis:
      type: object
      required: [rawTitle, channelName, thumbnailUrl, youtubeUrl, suggestedTitle, suggestedArtist, similarSongs]
      properties:
        rawTitle:
          type: string
        channelName:
          type: string
        thumbnailUrl:
          type: string
        youtubeUrl:
          type: string
        suggestedTitle:
          type: string
        suggestedArtist:
          type: string
        similarSongs:
          type: array
          items:
            $ref: "#/components/schemas/Song"
...
    GoogleSheetImport:
      type: object
      required: [sheetUrl]
      properties:
        sheetUrl:
          type: string
        sheetName:
          type: ["string", "null"]
    SyncPreview:
      type: object
      required: [rowsDetected, newSongs, updatedSongs, skippedSongs]
      properties:
        rowsDetected:
          type: integer
        newSongs:
          type: integer
        updatedSongs:
          type: integer
        skippedSongs:
          type: integer
        errors:
          type: array
          items:
            type: string
    ImportResult:
      type: object
      required: [rowsDetected, imported, updated, skipped, errors]
      properties:
        rowsDetected:
          type: integer
        imported:
          type: integer
        updated:
          type: integer
        skipped:
          type: integer
        errors:
          type: array
          items:
            type: string