# lib/api-zod/src/generated/types/cooldownStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 聆櫻聖境點歌系統 API
 * OpenAPI spec version: 0.1.0
 */
import type { CooldownStatusCooldownMode } from './cooldownStatusCooldownMode';

export interface CooldownStatus {
  cooldownDays: number;
  cooldownMode: CooldownStatusCooldownMode;
  /** @nullable */
  lastPerformed?: string | null;
  remainingDays?: number;
  isInCooldown: boolean;
}