# lib/api-zod/src/generated/types/songInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 聆櫻聖境點歌系統 API
 * OpenAPI spec version: 0.1.0
 */

export interface SongInput {
  title: string;
  artist: string;
  language: string;
  status?: string;
  /** @nullable */
  primaryTag?: string | null;
  categories?: string[];
  /** @nullable */
  youtubeUrl?: string | null;
  isPracticing?: boolean;
  hasPitchWarning?: boolean;
  cooldownDays?: number;
  cooldownMode?: string;
}