# lib/api-zod/src/generated/types/songUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 聆櫻聖境點歌系統 API
 * OpenAPI spec version: 0.1.0
 */

export interface SongUpdate {
  title?: string;
  artist?: string;
  language?: string;
  status?: string;
  /** @nullable */
  primaryTag?: string | null;
  categories?: string[];
  /** @nullable */
  youtubeUrl?: string | null;
  isPracticing?: boolean;
  hasPitchWarning?: boolean;
  cooldownDays?: number;
  cooldownMode?: string;
}